+++ src/ui/index.js
@@ -1,2 +1,2 @@
-export * from "./Footer/Footer";
-export * from "./Navbar/Navbar";
+export * from './Footer';
+export * from './Navbar';
